src/main/java/com/tw/apistack/endpoint/hello/HelloResource.java
package com.tw.apistack.endpoint.hello;

import java.util.concurrent.atomic.AtomicLong;

import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import com.tw.apistack.endpoint.hello.dto.Greeting;

/**
 * Created by jxzhong on 2017/6/27.
 */

@RestController
public class HelloResource {

    private static final String TEMPLATE = "Hello, %s!";
    private final AtomicLong counter = new AtomicLong();

    @GetMapping("/greeting")
    public Greeting greeting(@RequestParam(value = "name", defaultValue = "World") String name) {
        return new Greeting(counter.incrementAndGet(),
                String.format(TEMPLATE, name));
    }
}
